# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Blog Service","description":"Service to perform CRUD operations using goa","version":""},"host":"localhost:8000","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/":{"get":{"tags":["blog"],"summary":"list blog","description":"List all entries","operationId":"blog#list","responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/StoredblogResponse"}}}},"schemes":["http"]},"post":{"tags":["blog"],"summary":"create blog","description":"Add new blog and return its ID.","operationId":"blog#create","parameters":[{"name":"CreateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BlogCreateRequestBody"}}],"responses":{"201":{"description":"Created response.","schema":{"$ref":"#/definitions/BlogCreateResponseBody"}}},"schemes":["http"]}},"/oauth":{"post":{"tags":["blog"],"summary":"jwt blog","description":"Getting auth","operationId":"blog#jwt","parameters":[{"name":"X-Authorization","in":"header","description":"Access github token","required":false,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}}},"schemes":["http"]}},"/oauth/redirect":{"post":{"tags":["blog"],"summary":"oauth blog","description":"Github authentication to post a new blog","operationId":"blog#oauth","parameters":[{"name":"OauthRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BlogOauthRequestBody"}}],"responses":{"201":{"description":"Created response.","schema":{"type":"string"}}},"schemes":["http"]}},"/openapi.json":{"get":{"tags":["blog"],"summary":"Download ./gen/http/openapi.json","operationId":"blog#/openapi.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}},"/{id}":{"get":{"tags":["blog"],"summary":"show blog","description":"Show blog based on the id given","operationId":"blog#show","parameters":[{"name":"id","in":"path","description":"ID of a person","required":true,"type":"integer"},{"name":"ShowRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BlogShowRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/BlogShowResponseBody"}}},"schemes":["http"]},"delete":{"tags":["blog"],"summary":"remove blog","description":"Remove blog from storage","operationId":"blog#remove","parameters":[{"name":"id","in":"path","description":"ID of blog to remove","required":true,"type":"integer"}],"responses":{"204":{"description":"No Content response."}},"schemes":["http"]},"patch":{"tags":["blog"],"summary":"update blog","description":"Updating the existing blog","operationId":"blog#update","parameters":[{"name":"id","in":"path","description":"ID of blog to be updated","required":true,"type":"integer"},{"name":"UpdateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BlogUpdateRequestBody","required":["name","comments"]}}],"responses":{"204":{"description":"No Content response."}},"schemes":["http"]}},"/{id}/comments":{"post":{"tags":["blog"],"summary":"add blog","description":"Add new blog and return its ID.","operationId":"blog#add","parameters":[{"name":"id","in":"path","description":"Id of blog","required":true,"type":"integer"},{"name":"AddRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BlogAddRequestBody"}}],"responses":{"201":{"description":"Created response.","schema":{"$ref":"#/definitions/BlogAddResponseBody"}}},"schemes":["http"]}}},"definitions":{"BlogAddRequestBody":{"title":"BlogAddRequestBody","type":"object","properties":{"comments":{"$ref":"#/definitions/CommentsRequestBody"}},"example":{"comments":{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710}}},"BlogAddResponseBody":{"title":"BlogAddResponseBody","type":"object","properties":{"comments":{"$ref":"#/definitions/CommentsResponseBody"},"id":{"type":"integer","description":"Id of blog","example":3895848622,"format":"int32"}},"example":{"comments":{"comments":"Non eaque omnis.","id":3809683775},"id":544309705}},"BlogCreateRequestBody":{"title":"BlogCreateRequestBody","type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/CommentsRequestBody"},"description":"Comments","example":[{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710}],"maxItems":100},"id":{"type":"integer","description":"ID of a person","example":2926168972,"format":"int32"},"name":{"type":"string","description":"Name of person","example":"3wh","maxLength":100}},"example":{"comments":[{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710}],"id":1930843253,"name":"51m"}},"BlogCreateResponseBody":{"title":"BlogCreateResponseBody","type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/CommentsResponseBody"},"description":"Comments","example":[{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775}],"maxItems":100},"id":{"type":"integer","description":"ID of a person","example":3327564444,"format":"int32"},"name":{"type":"string","description":"Name of person","example":"m55","maxLength":100}},"example":{"comments":[{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775}],"id":998969093,"name":"obn"}},"BlogOauthRequestBody":{"title":"BlogOauthRequestBody","type":"object","properties":{"token":{"type":"string","description":"Access github token","example":"Est quas nam possimus asperiores."}},"example":{"token":"Doloribus recusandae cumque quia aut explicabo et."}},"BlogShowRequestBody":{"title":"BlogShowRequestBody","type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/CommentsRequestBody"},"description":"Comments","example":[{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710}],"maxItems":100},"name":{"type":"string","description":"Name of person","example":"uui","maxLength":100}},"example":{"comments":[{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710}],"name":"wlq"}},"BlogShowResponseBody":{"title":"BlogShowResponseBody","type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/CommentsResponseBody"},"description":"Comments","example":[{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775}],"maxItems":100},"id":{"type":"integer","description":"ID of a person","example":1177781145,"format":"int32"},"name":{"type":"string","description":"Name of person","example":"l1l","maxLength":100}},"example":{"comments":[{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775},{"comments":"Non eaque omnis.","id":3809683775}],"id":1824707723,"name":"pnl"}},"BlogUpdateRequestBody":{"title":"BlogUpdateRequestBody","type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/CommentsRequestBody"},"description":"Comments to be updated","example":[{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710}]},"name":{"type":"string","description":"Details of blog to be updated","example":"Cumque quo."}},"example":{"comments":[{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710},{"comments":"Et et incidunt cum omnis eligendi.","id":3140786710}],"name":"Odit dicta ex enim."},"required":["name","comments"]},"CommentsRequestBody":{"title":"CommentsRequestBody","type":"object","properties":{"comments":{"type":"string","description":"Comment for the blog","example":"Et et nobis est omnis consequatur rerum."},"id":{"type":"integer","description":"ID of a comment","example":2753009699,"format":"int32"}},"description":"Id and comments","example":{"comments":"Id aut assumenda et.","id":3343860529}},"CommentsResponse":{"title":"CommentsResponse","type":"object","properties":{"comments":{"type":"string","description":"Comment for the blog","example":"Quam quis veniam et quas nesciunt."},"id":{"type":"integer","description":"ID of a comment","example":1157799891,"format":"int32"}},"description":"Id and comments","example":{"comments":"Quasi deleniti ut.","id":35240643}},"CommentsResponseBody":{"title":"CommentsResponseBody","type":"object","properties":{"comments":{"type":"string","description":"Comment for the blog","example":"Aut tempore dolore."},"id":{"type":"integer","description":"ID of a comment","example":4012608446,"format":"int32"}},"description":"Id and comments","example":{"comments":"Perferendis dolores temporibus labore.","id":2497136214}},"StoredblogResponse":{"title":"Mediatype identifier: application/vnd.cellar.stored-blog; view=default","type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/CommentsResponse"},"description":"Comments","example":[{"comments":"Aperiam id corporis voluptatibus inventore.","id":2468836870},{"comments":"Aperiam id corporis voluptatibus inventore.","id":2468836870},{"comments":"Aperiam id corporis voluptatibus inventore.","id":2468836870}],"maxItems":100},"id":{"type":"integer","description":"ID is the unique id of the blog.","example":1983428894,"format":"int32"},"name":{"type":"string","description":"Name of person","example":"w6k","maxLength":100}},"description":"A Storedblog describes a blog retrieved by the storage service. (default view)","example":{"comments":[{"comments":"Aperiam id corporis voluptatibus inventore.","id":2468836870},{"comments":"Aperiam id corporis voluptatibus inventore.","id":2468836870},{"comments":"Aperiam id corporis voluptatibus inventore.","id":2468836870}],"id":652001506,"name":"6jw"},"required":["id","name"]}}}